# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TurbineOps Lite REST API
  version: 0.1.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /repair-plans/by-inspection/{inspectionId}:
    get:
      summary: Get repair plan by inspection id (VIEWER+). PostgreSQL is authoritative.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RepairPlanInspectionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairPlan'
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
        '404': { description: No repair plan for this inspection }
  /repair-plans/generate:
    post:
      summary: Generate or refresh repair plan from findings (ENGINEER+). Emits SSE `repair_plan_generated`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepairPlanGenerateRest'
      responses:
        '201':
          description: Created or updated (upsert per inspection)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairPlan'
        '400': { description: Validation error }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
        '404': { description: Inspection not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RepairPlanInspectionId:
      name: inspectionId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    RepairPlan:
      type: object
      properties:
        id: { type: string }
        inspectionId: { type: string }
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        totalEstimatedCost: { type: number }
        snapshotJson:
          type: object
          description: Versioned snapshot — summary plus serialized findings (PostgreSQL source of truth)
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RepairPlanGenerateRest:
      type: object
      required: [inspection_id]
      properties:
        inspection_id: { type: string }